+++ libs/common/src/main/java/org/opensearch/common/crypto/CryptoHandler.java
@@ -8,6 +8,7 @@
 
 package org.opensearch.common.crypto;
 
+import org.opensearch.common.annotation.ExperimentalApi;
 import org.opensearch.common.io.InputStreamContainer;
 
 import java.io.IOException;
@@ -20,6 +21,7 @@ import java.io.InputStream;
  * T - Encryption Metadata / CryptoContext
  * U - Parsed Encryption Metadata / CryptoContext
  */
+@ExperimentalApi
 public interface CryptoHandler<T, U> {
 
     /**